app/requirements.txt
opencv-python-headless>=4.9.0
mss>=9.0.1
requests>=2.31.0
pyyaml>=6.0.1
psutil>=5.9.8
pystray>=0.19.5
Pillow>=10.4.0
pywin32>=306
wmi>=1.5.1
python-dotenv>=1.0.0
